+++ ice/src/main/java/com/altinity/ice/cli/internal/cmd/InsertWatch.java
@@ -120,7 +120,14 @@ public class InsertWatch {
             logger.info("Inserting {}", insertBatch);
 
             try {
-              Insert.run(catalog, nsTable, insertBatch.toArray(String[]::new), options);
+              Insert.Result result =
+                  Insert.run(catalog, nsTable, insertBatch.toArray(String[]::new), options);
+              if (!result.ok()) {
+                logger.warn(
+                    "{}/{} file(s) failed to insert in this batch",
+                    result.totalNumberOfFiles(),
+                    result.numberOfFilesFailedToInsert());
+              }
             } catch (NoSuchTableException e) {
               if (!createTableIfNotExists) {
                 throw e;
@@ -145,7 +152,14 @@ public class InsertWatch {
                 retryInsert = false;
               }
               if (retryInsert) {
-                Insert.run(catalog, nsTable, insertBatch.toArray(String[]::new), options);
+                Insert.Result result =
+                    Insert.run(catalog, nsTable, insertBatch.toArray(String[]::new), options);
+                if (!result.ok()) {
+                  logger.warn(
+                      "{}/{} file(s) failed to insert in this batch",
+                      result.totalNumberOfFiles(),
+                      result.numberOfFilesFailedToInsert());
+                }
               }
             }
           }